Update README.md
Vibespace is an innovative social media platform that prioritizes creative expression. It allows users to share multimedia content, interact with others through likes, comments, and real-time notifications, and personalize their user profiles. The idea behind Vibespace is to provide a platform that merges the simplicity of social networking with powerful creative tools, making it an exciting space for users to showcase their individuality.